feat: refine user deletion logic and cleanup related entities

+++ src/main/java/com/friendhub/service/impl/UserServiceImpl.java
@@ -5,10 +5,14 @@ import com.friendhub.dto.request.UserUpdateRequest;
 import com.friendhub.dto.response.UserResponse;
 import com.friendhub.entity.Role;
 import com.friendhub.entity.User;
+import com.friendhub.enums.ErrorCode;
+import com.friendhub.enums.UserRole;
+import com.friendhub.exception.AppException;
 import com.friendhub.mapper.UserMapper;
 import com.friendhub.repository.RoleRepository;
 import com.friendhub.repository.UserRepository;
 import com.friendhub.service.UserService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,6 +20,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Service
+@RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
@@ -26,27 +31,17 @@ public class UserServiceImpl implements UserService {
 
     private final PasswordEncoder passwordEncoder;
 
-    public UserServiceImpl(UserRepository userRepository,
-                           RoleRepository roleRepository,
-                           UserMapper userMapper,
-                           PasswordEncoder passwordEncoder) {
-        this.userRepository = userRepository;
-        this.roleRepository = roleRepository;
-        this.userMapper = userMapper;
-        this.passwordEncoder = passwordEncoder;
-    }
-
     @Override
     public UserResponse createUser(UserCreationRequest request) {
         if (userRepository.existsByEmail(request.getEmail()))
-            throw new RuntimeException("User already existed.");
+            throw new AppException(ErrorCode.USER_ALREADY_EXISTED);
 
         User user = userMapper.toUser(request);
 
         user.setPassword(passwordEncoder.encode(request.getPassword()));
 
-        Role role = roleRepository.findById(2L)
-                        .orElseThrow(() -> new RuntimeException("Role not found"));
+        Role role = roleRepository.findByName(UserRole.MEMBER)
+                        .orElseThrow(() -> new AppException(ErrorCode.ROLE_NOT_FOUND));
 
         user.setRole(role);
 
@@ -62,26 +57,25 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserResponse getUserById(long userId) {
         return userMapper.toUserResponse(userRepository.findById(userId)
-                .orElseThrow(() -> new RuntimeException("User not found.")));
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND)));
     }
 
     @Override
     @Transactional
     public UserResponse updateUser(long id, UserUpdateRequest request) {
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found."));
+                .orElseThrow(() -> new AppException(ErrorCode.USER_NOT_FOUND));
 
         userMapper.updateUser(user, request);
-
-        userRepository.save(user);
 
         return userMapper.toUserResponse(user);
     }
 
     @Override
     public void deleteUser(long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new RuntimeException("User not found."));
+        boolean isExisted = userRepository.existsById(userId);
+        if (!isExisted)
+            throw new AppException(ErrorCode.USER_NOT_FOUND);
 
         userRepository.deleteById(userId);
     }
